Changed authentication and provided a different version of send command
Capture the tv_challenge in addition to the cookie. The challenge
number doesn't expire. So, you aren't constantly refreshing.

Changed some of the echo commands to work in Cygwin. The \n stuff
wasn't working very well.

Provided a different version of the send_command script that looks
up the code from its mnemonic in the ircc codes file.

The new script lso uses the tv_challenge value instead of the cookie
for authentication.

Added the file holding the tv_challenge code to the .gitignore file.